Ported to DE0 Nano, add new assembler and linker
This is large commit. Whole source is ported into DE0 Nano, this is
finall board to run MARK-II. There is example project for Quartus
13.0 sp1. Few examples in assembler are also added. VGA driver now
support more collors. Exsram was deleted, there is small ram from
M9K block that Cyclone IV have. This RAM is temporary and will be
replaced with SDRAM.